requirements.txt
discord.py 
mysql-connector-python 